Updates challenge (#1)
* Updates server

* Adds 2018 version

* Fixes figma link

* Removes yarn log

* Updates readme

* Refactors server

* Removes jquery

* Improves readme

* Adds unparrot

* Fixes nested list

* Fixes typo

* Refactors front end

* Refactors server
