--- back-end/users.json
@@ -1,62 +1,0 @@
-[
-  {
-    "id": "laverne_jacobi11",
-    "first_name": "Primary",
-    "last_name": "Christelle",
-    "avatar": "https://s3.amazonaws.com/uifaces/faces/twitter/romanbulah/128.jpg"
-  },
-  {
-    "id": "noelia_christiansen",
-    "first_name": "Virtual",
-    "last_name": "Emory",
-    "avatar": "https://s3.amazonaws.com/uifaces/faces/twitter/imcoding/128.jpg"
-  },
-  {
-    "id": "kevin_heaney",
-    "first_name": "Cross-Platform",
-    "last_name": "Veda",
-    "avatar": "https://s3.amazonaws.com/uifaces/faces/twitter/_victa/128.jpg"
-  },
-  {
-    "id": "ulises.rath",
-    "first_name": "Haptic",
-    "last_name": "Merle",
-    "avatar": "https://s3.amazonaws.com/uifaces/faces/twitter/peter576/128.jpg"
-  },
-  {
-    "id": "mona_mueller",
-    "first_name": "Wireless",
-    "last_name": "Michel",
-    "avatar": "https://s3.amazonaws.com/uifaces/faces/twitter/gonzalorobaina/128.jpg"
-  },
-  {
-    "id": "elinor.klein17",
-    "first_name": "Auxiliary",
-    "last_name": "Berry",
-    "avatar": "https://s3.amazonaws.com/uifaces/faces/twitter/shanehudson/128.jpg"
-  },
-  {
-    "id": "gunnar_gerhold",
-    "first_name": "Wireless",
-    "last_name": "Savanna",
-    "avatar": "https://s3.amazonaws.com/uifaces/faces/twitter/sydlawrence/128.jpg"
-  },
-  {
-    "id": "ramona_davis74",
-    "first_name": "Virtual",
-    "last_name": "Cordelia",
-    "avatar": "https://s3.amazonaws.com/uifaces/faces/twitter/santi_urso/128.jpg"
-  },
-  {
-    "id": "gerald47",
-    "first_name": "Digital",
-    "last_name": "Jaclyn",
-    "avatar": "https://s3.amazonaws.com/uifaces/faces/twitter/joshhemsley/128.jpg"
-  },
-  {
-    "id": "kieran56",
-    "first_name": "Mobile",
-    "last_name": "Hettie",
-    "avatar": "https://s3.amazonaws.com/uifaces/faces/twitter/_victa/128.jpg"
-  }
-]
