--- front-end/layout/README.md
@@ -1,5 +1,0 @@
-### Layout
-![Preview](layout.png)
-Desenhamos esse layout para que que ele sirva apenas de base para a sua implementação, não exigiremos que o resultado final esteja idêntico.
-
-Você pode encontrar o arquivo original em diversas extensões na pasta [`resources`](resources). A fonte usada em todos componentes é a [Lato](https://fonts.google.com/specimen/Lato) e os ícones que você vai precisar estão na pasta [`assets`](assets).
