+++ package.json
@@ -2,7 +2,7 @@
   "name": "TagChatter",
   "private": true,
   "version": "1.0.0",
-  "main": "index.js",
+  "main": "server.js",
   "license": "MIT",
   "dependencies": {
     "body-parser": "^1.17.2",
@@ -18,7 +18,7 @@
     "uuid": "^3.1.0"
   },
   "scripts": {
-    "start": "node index.js"
+    "start": "node server.js"
   },
   "engines": {
     "node": "~7"
