+++ src/index.html
@@ -1,0 +1,30 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <title>TagChatter</title>
+
+    <link rel="stylesheet" type="text/css" href="reset.css">
+    <link href="https://fonts.googleapis.com/css?family=IBM+Plex+Sans|Lato" rel="stylesheet">
+    <link rel="stylesheet" type="text/css" href="style.css">
+
+    <script src="index.js"></script>
+  </head>
+
+  <body class="wrapper">
+    <aside class="sidebar">
+      <div class="sidebar__logo">
+        <img src="images/logo.png" />
+      </div>
+    </aside>
+
+    <div class="content">
+      <header class="header">
+        <h1 class="header__title">#tagchatter</h1>
+
+        <div class="header__parrots-count">
+          <span class="header__title" id="parrots-counter">-</span>
+        </div>
+      </header>
+    </div>
+  </body>
+</html>
